+++ album-viewer/lang/translations.json
@@ -1,0 +1,10 @@
+[
+    {
+        "language": "en",
+        "values": {
+            "main-title": "Welcome to the world of the future",
+            "main-subtitle": "The future is now with copilot",
+            "main-button": "Get started"
+        }
+    }
+]
