+++ artists-api/Services/AlbumService.cs
@@ -1,0 +1,6 @@
+﻿namespace album_api.Services
+{
+    public class AlbumService
+    {
+    }
+}
