+++ artists-api/album-api.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33502.453
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artists-api", "artists-api.csproj", "{D2A481EA-A0FD-40A3-AD68-5056C27F4851}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "albums-api", "..\albums-api\albums-api.csproj", "{2B973170-1BC2-4578-8304-6814ECBD61B7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D2A481EA-A0FD-40A3-AD68-5056C27F4851}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D2A481EA-A0FD-40A3-AD68-5056C27F4851}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D2A481EA-A0FD-40A3-AD68-5056C27F4851}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D2A481EA-A0FD-40A3-AD68-5056C27F4851}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2B973170-1BC2-4578-8304-6814ECBD61B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B973170-1BC2-4578-8304-6814ECBD61B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B973170-1BC2-4578-8304-6814ECBD61B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B973170-1BC2-4578-8304-6814ECBD61B7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C1D74B1A-770A-4900-A8D4-C8AE54B6E790}
+	EndGlobalSection
+EndGlobal
